--- src/test/java/UItests/ApplicationManager.java
@@ -1,77 +1,0 @@
-package UItests;
-
-import dataObjects.newContactData;
-import dataObjects.newGroupData;
-import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.chrome.ChromeDriver;
-
-public class ApplicationManager {
-    String expected_url = "http://localhost:8080/addressbook/";
-    private WebDriver driver;
-
-    public void init() {
-        driver = new ChromeDriver();
-        driver.get(expected_url);
-        login("admin", "secret");
-    }
-
-    public void login(String account, String password) {
-        driver.manage().window().maximize();
-        driver.findElement(By.name("user")).sendKeys(account);
-        driver.findElement(By.name("pass")).click();
-        driver.findElement(By.name("pass")).sendKeys(password);
-        driver.findElement(By.cssSelector("input:nth-child(7)")).click();
-    }
-
-    public void stop() {
-        if (driver != null) {
-            driver.quit();
-        }
-    }
-
-    public void clickSubmitBtn() {
-        driver.findElement(By.name("submit")).click();
-    }
-
-    public void fillInRequiredNewGroupFields(newGroupData newGroupData) {
-        driver.findElement(By.name("group_name")).click();
-        driver.findElement(By.name("group_name")).sendKeys(newGroupData.getGroupName());
-        driver.findElement(By.name("group_header")).click();
-        driver.findElement(By.name("group_header")).sendKeys(newGroupData.getGroupDescription1());
-        driver.findElement(By.name("group_footer")).click();
-        driver.findElement(By.name("group_footer")).sendKeys(newGroupData.getGroupDescription2());
-    }
-
-    public void selectCreateNewGroup() {
-        driver.findElement(By.name("new")).click();
-    }
-
-    public void openGroupPage() {
-        driver.findElement(By.linkText("groups")).click();
-    }
-
-    public void fillInNewContactFields(newContactData newContactData) {
-        driver.findElement(By.name("firstname")).sendKeys(newContactData.getFirstName());
-        driver.findElement(By.name("lastname")).sendKeys(newContactData.getLastName());
-        driver.findElement(By.name("nickname")).sendKeys(newContactData.getNickName());
-        driver.findElement(By.name("mobile")).sendKeys(newContactData.getMobile());
-        driver.findElement(By.name("email")).sendKeys(newContactData.getEmail());
-    }
-
-    public void openCreateNewContact() {
-        driver.findElement(By.xpath("//a[text()='add new']")).click();
-    }
-
-    public void return2GroupPage() {
-        driver.findElement(By.linkText("group page")).click();
-    }
-
-    public void clickDeleteGroupBtn() {
-        driver.findElement(By.name("delete")).click();
-    }
-
-    public void selectFirstGroupInList() {
-        driver.findElement(By.name("selected[]")).click();
-    }
-}
